tests/requirements_test.txt
# Test requirements for PyTorch functionality testing
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0

# Core dependencies for testing
numpy>=1.21.0
opencv-python>=4.6.0
Pillow>=9.0.0

# PyTorch testing dependencies
torch>=1.12.0
torchvision>=0.13.0
segmentation-models-pytorch>=0.3.0
albumentations>=1.3.0

# TensorFlow testing dependencies (optional)
tensorflow>=2.10.0
segmentation-models>=1.0.1

# PyQt5 testing (mock-heavy)
PyQt5>=5.15.0

# Additional testing utilities
pyyaml>=6.0